requirements.txt
torch==2.0.0
torchvision==0.15.0
numpy==1.24.3
scikit-learn==1.0.2
matplotlib==3.7.1
seaborn==0.12.2
pillow==9.5.0
pandas==2.0.3
streamlit==1.28.0
tqdm==4.65.0 